debundle/lowering: extract anonymous-statement matcher
BAZEL_TEST_INVOCATIONS=none: file move, no semantic change.
